+++ Service/Implement/ChannelService.cs
@@ -92,7 +92,7 @@ namespace Service
 
             // Cập nhật thuộc tính của đối tượng channel
             channel.UserName = id;
-            channel.Platform = (int)ePlatform;
+            channel.Platform = ePlatform;
             _mapper.Map(data, channel);
 
             return channel;
